pyproject.toml
[project]
name = "ndtp-bot"
version = "0.1.0"
description = ""
authors = [
    {name = "LightMizu",email = "1blend.your.life1@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiogram (==3.15.0)",
    "aiohttp (>=3.9.0)",
    "python-dotenv (==1.0.0)",
    "pydantic (>=2.0.0)",
    "sentence-transformers (==2.3.1)",
    "chromadb (==0.4.22)",
    "numpy (>=1.24.0)",
    "scikit-learn (>=1.3.0)",
    "transformers (>4.30.0)",
    "beautifulsoup4 (>=4.12.0)",
    "requests (>=2.31.0)",
    "lxml (>=4.9.0)",
    "pypdf2 (>=3.0.0)",
    "pdf2image (>=1.17.0)",
    "pytesseract (>=0.3.10)",
    "pillow (>=10.0.0)",
    "rich (>=13.0.0)",
    "dateparser (>=1.1.8)",
    "redis (>=5)",
    "tenacity (>=8.2.0)",
    "torch (==2.7.1)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "boto3 (>=1.40.19,<2.0.0)",
    "sqlalchemy (>=2.0.43,<3.0.0)",
    "pymupdf (>=1.26.4,<2.0.0)",
    "psycopg2 (>=2.9.10,<3.0.0)"
]

[[tool.poetry.source]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
priority = "explicit"



[tool.poetry.dependencies]
torch = {source = "pytorch-cpu"}

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
